# Improved performance of django.forms.ChoiceWidget.optgroups()