@carltongibson 2021-01-27
Hey @smithdc1, thanks. I think ultimately readability trumps unless we can show that something is on the critical path in realistic circumstances. (Something in URL resolution, for example 😃, is worth leaning more towards speed.) 

If you want to reduce this down to just the Boolean change, we’d be happy with that. 

👍